+++ UVtools.Core/FileFormats/VDAFile.cs
@@ -153,6 +153,8 @@ public class VDAFile : FileFormat
 
     public override FileFormatType FileType => FileFormatType.Archive;
 
+    public override string ConvertMenuGroup => "Voxeldance";
+
     public override FileExtension[] FileExtensions { get; } = {
         new(typeof(VDAFile), "zip", "Voxeldance Additive Zip")
     };