+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 18/09/2022 - v3.6.5
+
+- **UI:** 
+  - (Add) Groups to convert menu
+  - (Improvement) Re-arrange convert menu order
+  - (Fix) Settings: Open file dialog filter dropdown width
+- **PrusaSlicer:**
+   - (Add) AnyCubic Photon D2
+   - (Add) Creality Halot Ray CL925
+   - (Add) Elegoo Saturn 8K
+   - (Add) Uniformation GKtwo
+   - (Add) Nova3D Whale3 Pro
+   - (Add) Phrozen Sonic Mighty 8K
+- (Add) AnyCubic Photon D2 (.dl2p) compatibility to AnyCubic file format
+- (Fix) Hide "Debug tools" from "Help" menu
+
 ## 15/09/2022 - v3.6.4
 
 - **Windows MSI installer:**